Moved AlertModal from Identify into shared
