+++ app/webpack/observations/identify/actions/identification_actions.js
@@ -8,7 +8,7 @@ import {
   addIdentification
 } from "./current_observation_actions";
 import { updateObservationInCollection } from "./observations_actions";
-import { showAlert } from "./alert_actions";
+import { showAlert } from "../../../shared/ducks/alert_modal";
 
 const POST_IDENTIFICATION = "post_identification";
 const AGREEING_WITH_OBSERVATION = "agreeing_with_observation";
