+++ app/webpack/observations/identify/actions/index.js
@@ -6,4 +6,3 @@ export * from "./observations_actions";
 export * from "./observations_stats_actions";
 export * from "./search_params_actions";
 export * from "./finished_modal_actions";
-export * from "./alert_actions";
