+++ app/webpack/observations/identify/reducers/index.js
@@ -32,7 +32,7 @@ import currentObservation from "./current_observation_reducer";
 import config from "../../../shared/ducks/config";
 import searchParams from "./search_params_reducer";
 import finishedModal from "./finished_modal_reducer";
-import alert from "./alert_reducer";
+import alertModal from "../../../shared/ducks/alert_modal";
 import textEditor from "../../shared/ducks/text_editors";
 import suggestions from "../ducks/suggestions";
 import controlledTerms from "../../show/ducks/controlled_terms";
@@ -50,7 +50,7 @@ const rootReducer = combineReducers( {
   currentObservation,
   searchParams,
   finishedModal,
-  alert,
+  alertModal,
   textEditor,
   suggestions,
   controlledTerms,
