+++ app/webpack/shared/components/alert_modal.jsx
